Refactor FileUtils
