+++ app/src/main/java/swati4star/createpdf/util/ExtractImages.java
@@ -16,7 +16,7 @@ import java.util.ArrayList;
 import swati4star.createpdf.interfaces.ExtractImagesListener;
 
 import static swati4star.createpdf.util.FileUtils.getFileNameWithoutExtension;
-import static swati4star.createpdf.util.FileUtils.saveImage;
+import static swati4star.createpdf.util.ImageUtils.saveImage;
 
 public class ExtractImages extends AsyncTask<Void, Void, Void> {
     private final String mPath;
